// src/models/HorizonViewDatacenter.ts
/* WARNING: This file is auto-generated . Do not edit manually. */

export type HorizonViewDatacenter = {
  activeDirectoryIP: string;
  adminDomain: string;
  adminNetworkNextHop: string;
  adminPrivateNetwork: string;
  customerIntercoIP: string;
  customerIntercoMask: string;
  customerIntercoVlanId: number;
  datacenterId: number;
  name: string;
  serviceName: string;
};